Clean labels and module names.
