+++ lib/screens/projects/components/menu_drawer.dart
@@ -4,7 +4,7 @@ import 'package:nahpu/services/providers/projects.dart';
 import 'package:nahpu/screens/export/bundle_project.dart';
 import 'package:nahpu/screens/export/export_db.dart';
 import 'package:nahpu/screens/export/export_settings.dart';
-import 'package:nahpu/screens/export/export_pdf.dart';
+import 'package:nahpu/screens/export/export_documents.dart';
 import 'package:nahpu/screens/export/export_records.dart';
 import 'package:nahpu/screens/export/export_report.dart';
 import 'package:nahpu/screens/projects/new_project.dart';
@@ -67,22 +67,22 @@ class ProjectMenuDrawerState extends ConsumerState<ProjectMenuDrawer> {
           },
         ),
         ListTile(
-          leading: const Icon(Icons.picture_as_pdf_rounded),
-          title: const Text('Export to pdf'),
-          onTap: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => const ExportPdfForm()),
-            );
-          },
-        ),
-        ListTile(
           leading: Icon(Icons.adaptive.share_rounded),
           title: const Text('Export records'),
           onTap: () {
             Navigator.push(
               context,
               MaterialPageRoute(builder: (context) => const ExportForm()),
+            );
+          },
+        ),
+        ListTile(
+          leading: const Icon(Icons.picture_as_pdf_rounded),
+          title: const Text('Export documents'),
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(builder: (context) => const ExportPdfForm()),
             );
           },
         ),
